feat: feature flags, update sha, change openapi to priv
